% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countCombinations.R
\name{countCombinations}
\alias{countCombinations}
\title{Calculate number of possible group combinations}
\usage{
countCombinations(
  object = NULL,
  replicate_labels = NULL,
  group_labels = NULL,
  use_cells = NULL,
  n_replicates = NULL,
  n_group1 = NULL
)
}
\arguments{
\item{object}{An object of class 'Seurat' or 'SingleCellExperiment'.}

\item{replicate_labels}{A character string or vector indicating the name of the
column containing the replicate labels. Defaults to \code{NULL}.}

\item{group_labels}{A character string or vector indicating the name of the
column containing the two comparison group labels. Defaults to \code{NULL}.}

\item{use_cells}{A vector of cell names subset to. Default = \code{NULL} will
use all cells.}

\item{n_replicates}{A numeric value indicating the total number of replicates.
Defaults to \code{NULL}.}

\item{n_group1}{A numeric value indicating the number of replicates that belong
to one of the two groups (doesn't matter which). Defaults to \code{NULL}.}
}
\value{
Returns a numeric value indicating the number of possible group
combinations.
}
\description{
For a set of replicates that belong to two groups, this function will
calculate how many different randomized group combinations are possible
(i.e., how many ways to divide the replicates into two groups while retaining
the same number of replicates per group). The input should be either the
replicate and group column names from an object's metadata, or simply the
total number of replicates (n) and the number of samples in one of the two
groups (k).
}
\details{
The formula is: C = n!/(k!(n-k)!)
(see base function \code{choose})
}
