% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCombinations.R
\name{getCombinations}
\alias{getCombinations}
\title{Generate combinations}
\usage{
getCombinations(
  object = NULL,
  replicate_labels = NULL,
  group_labels = NULL,
  use_cells = NULL,
  n_replicates = NULL,
  n_group1 = NULL,
  n_combinations = 1000,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class 'Seurat' or 'SingleCellExperiment'.}

\item{replicate_labels}{A character string or vector indicating the name of the
column containing the replicate labels. Defaults to \code{NULL}.}

\item{group_labels}{A character string or vector indicating the name of the
column containing the two comparison group labels. Defaults to \code{NULL}.}

\item{use_cells}{A vector of cell names subset to. Default = \code{NULL} will
use all cells.}

\item{n_replicates}{A numeric value indicating the total number of replicates.
Defaults to \code{NULL}.}

\item{n_group1}{A numeric value indicating the number of replicates that belong
to one of the two groups (doesn't matter which). Defaults to \code{NULL}.}

\item{n_combinations}{A numeric value indicating the number of combinations
to generate. Defaults to 1000.}

\item{verbose}{A boolean value indicating whether to use verbose output
during the execution of this function. Defaults to \code{TRUE}.
Can be set to \code{FALSE} for a cleaner output.}
}
\value{
Returns a matrix where each column contains a combination of index
values indicating which replicates to assign to the first group.
}
\description{
For a set of replicates that belong to two groups, this function will generate
a set of randomized group combinations. The input should be either the replicate
and group column names from an object's metadata, or simply the total number
of replicates (n) and the number of replicates in one of the two groups (k).
}
