% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingUtils.R
\name{getHistograms}
\alias{getHistograms}
\title{Generate histogram plots of permutation test results}
\usage{
getHistograms(input, title = NULL, subtitle = NULL, label_pvalue = TRUE)
}
\arguments{
\item{input}{Output from function \code{permuteDE}.}

\item{title}{Character string indicating the plot title. Default = \code{NULL}
sets a title automatically for each split.}

\item{subtitle}{Character string indicating the plot subtitle. Default =
\code{NULL} automatically generates a subtitle.}

\item{label_pvalue}{A Boolean value indicating whether to label the
permutation test p-value. Defaults to TRUE.}
}
\value{
A named list of \code{ggplot2} objects, where each element
corresponds to a volcano plot for one split of the DE results.
}
\description{
This function takes the output from function \code{permuteDE} and creates a
list containing a histogram plot showing the permutation test result for
each split. Each histogram includes a vertical line at the observed number
of DE features in the unpermuted "true" comparison.
}
