% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPseudobulk.R
\name{getPseudobulk}
\alias{getPseudobulk}
\title{Calculate pseudobulk values}
\usage{
getPseudobulk(
  object,
  replicate_labels,
  split_labels = NULL,
  use_cells = NULL,
  min_cells_per_split = 1,
  min_replicates_per_split = 1,
  min_cells_per_feature = 0,
  use_assay = NULL,
  use_layer = NULL,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{A 'Seurat' or 'SingleCellExperiment' object.}

\item{replicate_labels}{A character string or vector indicating the name of the
column containing the replicate labels.}

\item{split_labels}{A character string or vector indicating the name of a
column by which to split the cells prior to pseudobulking and performing
differential expression (e.g., cell types). Results will be returned for
each unique value in the column indicated by 'split_labels'. Default =
\code{NULL} will run pseudobulk differential expression on all cells
together.}

\item{use_cells}{A vector of cell names subset to. Default = \code{NULL} will
use all cells.}

\item{min_cells_per_split}{A numeric value indicating the minimum number of
cells within one split. Pseudobulk expression will not be calculated for
splits with fewer cells. Defaults to 1.}

\item{min_replicates_per_split}{A numeric value indicating the minimum number
of distinct replicates represented within one split. Pseudobulk expression
will not be calculated for splits with fewer replicates. Defaults to 1.}

\item{min_cells_per_feature}{A numeric value indicating the minimum number
of cells (within a split) with expression of a gene. Pseudobulk expression
will not be calculated for genes with fewer cells. Defaults to 0.}

\item{use_assay}{A character string indicating the assay to use in the
provided object. Default = \code{NULL} will choose the current active assay
for Seurat objects and the \code{counts} assay for SingleCellExperiment
objects.}

\item{use_layer}{For Seurat objects, a character string or vector indicating
the layer — previously known as slot — to use in the provided object.
Default = \code{NULL} will use the \code{counts} layer.}

\item{n_cores}{A numeric value indicating the number of cores to use for
parallelization. Default = \code{NULL} will use the number of available cores
minus 2.}

\item{verbose}{A boolean value indicating whether to use verbose output
during the execution of this function. Defaults to \code{TRUE}.
Can be set to \code{FALSE} for a cleaner output.}
}
\value{
Returns a list with one pseudobulk (gene x replicate) matrix per
split.
}
\description{
This function will generate a set of pseudobulk expression matrices.
}
\details{
This function was inspired by R package neurorestore/Libra
(Squair et al. 2021).
}
