% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperUtils.R
\name{name A string under which data are stored}
\alias{name A string under which data are stored}
\alias{.retrieveData}
\title{Retrieve data from an object}
\usage{
.retrieveData(object, type, name, use_cells = NULL)
}
\arguments{
\item{object}{A 'Seurat' or 'SingleCellExperiment' object}

\item{type}{A string indicating data type to retrieve}
}
\description{
Retrieve data from an object
}
