% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingUtils.R
\name{palette_permuteDE}
\alias{palette_permuteDE}
\title{Generate color palette}
\usage{
palette_permuteDE(type = "discrete", n = 100)
}
\arguments{
\item{type}{A character string indicating the palette type. Permitted values
are "discrete" and "gradient". Defaults to discrete.}

\item{n}{Number of colors, defaults to 100.}
}
\value{
Returns a vector of n hex values.
}
\description{
Generate a color palette. For discrete palettes, hex values are standard up
to n = 100, but for larger values of n, they are generated using
\code{Polychrome::createPalette()}.
}
\details{
Adapted from function  \code{CHOIR::CHOIRpalette} from R package \code{CHOIR}
(Sant et al. 2025).
}
