% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingUtils.R
\name{plotDimReduction}
\alias{plotDimReduction}
\title{Plot dimensionality reduction}
\usage{
plotDimReduction(
  reduction,
  input = NULL,
  split_labels = NULL,
  use_cells = NULL,
  color_by = NULL,
  permutation_test_alpha = 1,
  label_splits = FALSE,
  label_statistics = FALSE,
  palette_name = NULL,
  ...
)
}
\arguments{
\item{reduction}{Dimensionality reduction cell coordinates. If there are
more than two columns, only the first two will be used.}

\item{input}{Output from function \code{permuteDE}.}

\item{split_labels}{An character vector containing the split labels for each
cell in order. Default = \code{NULL} will assume all cells belong to the same
split.}

\item{use_cells}{A vector of cell names to subset the dimensionality
reduction cell coordinates to. Default = \code{NULL} will use all cells.}

\item{color_by}{A character vector indicating what metric to use to color
each split. Permitted values are "n_sig", "pvalue", and "split". Default =
\code{NULL} will not color the cells.}

\item{permutation_test_alpha}{A numeric value indicating the significance
level to apply to the permutation test results. Splits that do not pass this
threshold will be grayed out. Default = 1 applies no threshold.}

\item{label_splits}{A Boolean value indicating whether to label the
splits. Defaults to \code{FALSE}.}

\item{label_statistics}{A Boolean value indicating whether to label the
values of the selected metric. Defaults to \code{FALSE}.}

\item{palette_name}{A character string indicating the palette name. Permitted
values are "choir", "archr", "corces_cold", and "corces_warm". Default =
\code{NULL} will use "choir" for discrete colors and "corces_cold" for
gradient colors.}

\item{...}{Extra parameters passed to \code{Seurat::DimPlot()} or
\code{Seurat::FeaturePlot()}.}
}
\value{
A dimensionality reduction plot.
}
\description{
This function will generate a dimensionality reduction plot colored according
to the selected metric.
}
