% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingUtils.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Generate histogram plot(s) of permutation test results}
\usage{
plotHistogram(
  input,
  use_splits = NULL,
  title = NULL,
  subtitle = NULL,
  label_pvalue = TRUE
)
}
\arguments{
\item{input}{Output from function \code{permuteDE}.}

\item{use_splits}{A character string or vector containing the names of splits
to use. Defaults to \code{NULL}, which will try all splits.}

\item{title}{Character string indicating the plot title. Default = \code{NULL}
sets a title automatically for each split.}

\item{subtitle}{Character string indicating the plot subtitle. Default =
\code{NULL} automatically generates a subtitle.}

\item{label_pvalue}{A Boolean value indicating whether to label the
permutation test p-value. Defaults to TRUE.}
}
\value{
If only one split is provided, a single \code{ggplot2} object,
otherwise a named list of \code{ggplot2} objects, where each element
corresponds to a split.
}
\description{
This function takes the output from function \code{permuteDE} and creates a
histogram plot showing the permutation test result for each indicated split.
Each histogram includes a vertical line at the observed number of DE features
in the unpermuted "true" comparison.
}
