% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingUtils.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Generate volcano plot(s) from differential expression results}
\usage{
plotVolcano(
  input,
  alpha = 0.05,
  lfc_threshold = 0.5,
  use_splits = NULL,
  title = NULL,
  subtitle = NULL,
  n_max_label = 10,
  center = TRUE
)
}
\arguments{
\item{input}{Output from function \code{runDE} or a list containing
(at minimum) a dataframe named "DE_results" including columns:
\describe{
\item{gene}{Gene identifiers (character).}
\item{lfc}{Log2 fold change values (numeric).}
\item{padj}{Adjusted p-values (numeric).}
\item{split}{(Optionally) Grouping variable used to subset results
(character or factor).}
}}

\item{alpha}{A numeric value indicating the significance level used for
permutation test comparisons of the number of differentially expressed
features. Defaults to 0.05.}

\item{lfc_threshold}{A numeric value indicating the minimum absolute value
log fold change for a gene to be counted as a "hit". Defaults to 0.5. Set to
0 to disregard log fold change when counting hits.}

\item{use_splits}{A character string or vector containing the names of splits
to use. Defaults to \code{NULL}, which will try all splits.}

\item{title}{Character string indicating the plot title. Default = \code{NULL}
sets a title automatically for each split.}

\item{subtitle}{Character string indicating the plot subtitle. Default =
\code{NULL} automatically generates a subtitle describing the DE method used.}

\item{n_max_label}{A numeric value indicating how many of the top
significant DE features to label. Defaults to 10.}

\item{center}{A Boolean value indicating whether to center the x-axis at 0.
Defaults to \code{TRUE}.}
}
\value{
If only one split is provided, a single \code{ggplot2} object,
otherwise a named list of \code{ggplot2} objects, where each element
corresponds to a split.
}
\description{
This function takes the output from function \code{runDE} and creates volcano
plots for each indicated split.
}
