% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDE.R
\name{runDE}
\alias{runDE}
\title{Run pseudobulk differential expression}
\usage{
runDE(
  object,
  replicate_labels,
  group_labels,
  split_labels = NULL,
  force_balance = FALSE,
  reference_group = NULL,
  use_cells = NULL,
  min_cells_per_split = 4,
  min_replicates_per_split = 4,
  min_replicates_per_group = 2,
  min_cells_per_feature = 1,
  de_method = "edgeR",
  de_test = "LRT",
  p_adjust_method = "fdr",
  use_assay = NULL,
  use_layer = NULL,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class 'Seurat' or 'SingleCellExperiment'.}

\item{replicate_labels}{A character string or vector indicating the name of the
column containing the replicate labels.}

\item{group_labels}{A character string or vector indicating the name of the
column containing the two comparison group labels.}

\item{split_labels}{A character string or vector indicating the name of a
column by which to split the cells prior to pseudobulking and performing
differential expression (e.g., cell types). Results will be returned for
each unique value in the column indicated by 'split_labels'. Default =
\code{NULL} will run pseudobulk differential expression on all cells
together.}

\item{force_balance}{A boolean indicating if two groups have equal sample size.
Default to \code{FALSE}. If TRUE, and the two groups have unequal sample sizes,
the larger group will be randomly downsampled to match the size of the smaller group.}

\item{reference_group}{A string specifying the reference group. Default to
\code{NULL}, in which case the first value in the group column is used as the reference.}

\item{use_cells}{A vector of cell names subset to. Default = \code{NULL} will
use all cells.}

\item{min_cells_per_split}{A numeric value indicating the minimum number of
cells within one split. Pseudobulk expression and differential expression
will not be performed for splits with fewer cells. Defaults to 4.}

\item{min_replicates_per_split}{A numeric value indicating the minimum number
of distinct replicates represented within one split. Pseudobulk expression
and differential expression will not be performed for splits with fewer
replicates. Defaults to 4.}

\item{min_replicates_per_group}{A numeric value indicating the minimum number
of distinct replicates represented within each of the two comparison groups.
Pseudobulk expression and differential expression will not be performed for
splits with fewer replicates. Defaults to 2.}

\item{min_cells_per_feature}{A numeric value indicating the minimum number
of cells (within a split) with expression of a gene. Pseudobulk expression
and differential expression will not be calculated for genes with fewer
cells. Defaults to 1.}

\item{de_method}{Which tool to use for differential expression. Permitted
values are 'edgeR', 'DESeq2', and 'limma'. Defaults to 'edgeR'.}

\item{de_test}{Which test to use for differential expression. Defaults to
'LRT'.}

\item{p_adjust_method}{A string indicating which multiple comparison
adjustment to use. For permitted values, see \code{stats::p.adjust}. Defaults
to 'fdr' (Benjamini & Hochberg, 1995).}

\item{use_assay}{A character string indicating the assay to use in the
provided object. Default = \code{NULL} will choose the current active assay
for Seurat objects and the \code{counts} assay for SingleCellExperiment
objects.}

\item{use_layer}{For Seurat objects, a character string or vector indicating
the layer — previously known as slot — to use in the provided object.
Default = \code{NULL} will use the \code{counts} layer.}

\item{n_cores}{A numeric value indicating the number of cores to use for
parallelization. Default = \code{NULL} will use the number of available cores
minus 2.}

\item{verbose}{A boolean value indicating whether to use verbose output
during the execution of this function. Defaults to \code{TRUE}.
Can be set to \code{FALSE} for a cleaner output.}
}
\value{
Returns a list containing the following elements: \describe{
\item{DE_results}{Dataframe containing DE results for each feature, by
split}
\item{PB_values}{List of feature x replicate matri(ces) containing
pseudobulk values for each feature, one per split}
\item{group_key}{Dataframe record of group labels corresponding to each replicate}
\item{parameters}{Dataframe record of parameter values used}
}
}
\description{
Generate pseudobulk matri(ces) and run pseudobulk differential expression
between two groups for each using existing tools: \code{edgeR},
\code{DESeq2}, or \code{limma}.
}
\details{
This function was inspired by R package neurorestore/Libra and some aspects
are adapted therefrom (Squair et al. 2021).
}
