% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSCDE.R
\name{runSCDE}
\alias{runSCDE}
\title{Run single-cell differential expression between two groups using \code{Seurat} defaults.}
\usage{
runSCDE(
  object,
  group_labels,
  reference_group,
  force_balance = FALSE,
  use_cells = NULL,
  min_replicates_per_group = 2,
  min_cells_per_feature = 1,
  use_assay = NULL,
  use_layer = NULL,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class 'Seurat'.}

\item{group_labels}{A character string or vector indicating the name of the
column containing the two comparison group labels.}

\item{reference_group}{A string specifying the reference group. Default to
\code{NULL}, in which case the first value in the group column is used as the reference.}

\item{force_balance}{A boolean indicating if two groups have equal sample size.
Default to \code{FALSE}. If TRUE, and the two groups have unequal sample sizes,
the larger group will be randomly downsampled to match the size of the smaller group.}

\item{use_cells}{A vector of cell names subset to. Default = \code{NULL} will
use all cells.}

\item{min_replicates_per_group}{A numeric value indicating the minimum number
of distinct replicates represented within each of the two comparison groups.
Pseudobulk expression and differential expression will not be performed for
splits with fewer replicates. Defaults to 2.}

\item{min_cells_per_feature}{A numeric value indicating the minimum number
of cells (within a split) with expression of a gene. Pseudobulk expression
and differential expression will not be calculated for genes with fewer
cells. Defaults to 1.}

\item{use_assay}{A character string indicating the assay to use in the
provided object. Default = \code{NULL} will choose the \code{RNA} assay.}

\item{use_layer}{a character string or vector indicating
the layer — previously known as slot — to use in the provided object}

\item{n_cores}{A numeric value indicating the number of cores to use for
parallelization. Default = \code{NULL} will use the number of available cores
minus 2.}

\item{verbose}{A boolean value indicating whether to use verbose output
during the execution of this function. Defaults to \code{TRUE}.
Can be set to \code{FALSE} for a cleaner output.}
}
\value{
A data.frame of differentially expressed genes from \code{FindMarkers()}.
}
\description{
This function performs differential expression analysis between two
groups of cells in a Seurat object using Seurat's \code{FindMarkers()}.
}
