% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample dataset: sample_data}
\format{
A Seurat object with 2000 features across 1450 cells:
\describe{
\item{counts}{Count matrix}
\item{meta.data}{Metadata columns: sample, group, cell_type}
\item{umap}{UMAP coordinates}
}
}
\usage{
sample_data
}
\description{
A small Seurat object with sample values.
}
\keyword{datasets}
