% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countCombinations.R
\name{countCombinations}
\alias{countCombinations}
\title{Calculate the number of possible group combinations}
\usage{
countCombinations(
  object = NULL,
  replicate_labels = NULL,
  group_labels = NULL,
  use_cells = NULL,
  n_replicates = NULL,
  n_group1 = NULL
)
}
\arguments{
\item{object}{A 'Seurat' or 'SingleCellExperiment' object.}

\item{use_cells}{A vector of cell names to subset. Defaults to \code{NULL},
which includes all cells.}

\item{n_replicates}{A numeric value indicating the total number of replicates.
Defaults to \code{NULL}.}

\item{n_group1}{A numeric value indicating the number of replicates in one
group (doesn't matter which). Defaults to \code{NULL}.}

\item{replicate_label}{A string or character vector indicating the column
containing replicate labels. Defaults to \code{NULL}.}

\item{group_label}{A string or character vector indicating the column
containing two comparison group labels. Defaults to \code{NULL}.}
}
\value{
Returns a numeric value indicating the number of all possible group
combinations.
}
\description{
Given two groups of replicates, this function calculates the number of unique
ways to divide the replicates into two groups while preserving group sizes.
The formula is: C = n!/(k!(n-k)!). The input takes two forms: (1) replicate
and group column names; (2) total number of replicates (n) and the number of
replicates in one group (k)(doesn't matter which).
}
\details{
(see base function \code{choose})
}
