% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHistogram.R
\name{generateHistogram}
\alias{generateHistogram}
\title{Generate Histograms of Permutation Test Results}
\usage{
generateHistogram(permuteDE_result, title = NULL)
}
\arguments{
\item{permuteDE_result}{A returned list from \code{permuteDE}}

\item{title}{A string to be the title of histograms. Default to \code{NULL}.}
}
\description{
This function creates a list of \code{ggplot2} histogram plots showing the distribution
of the number of differentially expressed (DE) genes across permutations for each
split in the permutation DE results. Each histogram includes a vertical line at
the observed number of DE genes and a p-value annotation.
}
