% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCombinations.R
\name{getCombinations}
\alias{getCombinations}
\title{Generate combinations}
\usage{
getCombinations(
  object = NULL,
  replicate_labels = NULL,
  group_labels = NULL,
  use_cells = NULL,
  n_replicates = NULL,
  n_group1 = NULL,
  n_combinations = 1000,
  message = "",
  random_seed = 1,
  verbose = TRUE
)
}
\arguments{
\item{object}{An optional 'Seurat' or 'SingleCellExperiment' object. If
\code{NULL}, \code{countCombinations} will expect either (1) vector input to
parameters \code{replicate_label} and \code{group_label} or (2) numeric input
to parameters \code{n_replicates} and \code{n_group1}.}

\item{replicate_labels}{A string indicating the name of the
metadata column containing the biological replicate labels or a character
vector containing the biological replicate labels in order.}

\item{group_labels}{A string indicating the name of the
column containing the two comparison group labels or a character vector
containing the comparison labels in order.}

\item{use_cells}{A vector of cell names to subset prior to calculating
possible group comibinations. Default = \code{NULL} will use all cells.}

\item{n_replicates}{A numeric value indicating the total number of
replicates. Defaults to \code{NULL}.}

\item{n_group1}{A numeric value indicating the number of replicates in one
group (doesn't matter which). Defaults to \code{NULL}.}

\item{n_combinations}{A numeric value indicating the number of combinations
to generate. Defaults to 1000.}

\item{message}{A character string indicating additional progress messaging
(internal use). Defaults to "".}

\item{random_seed}{A numeric value indicating the random seed to be used.
Defaults to 1.}

\item{verbose}{A Boolean value indicating whether to use verbose output
during the execution of this function. Defaults to \code{TRUE}.
Can be set to \code{FALSE} for a cleaner output.}
}
\value{
Returns a matrix where each column contains a combination of index
values indicating which replicates to assign to the first group.
}
\description{
For a set of replicates that belong to two groups, this function will
generate a set of randomized group combinations.
}
\details{
Users may provide input in three ways:

(1) A vector of replicate labels and a vector of group labels (in order)

(2) The total number of replicates (n) and the number of
replicates in one group (k) (doesn't matter which group).

(3) Column names indicating replicate and group metadata columns in a
provided 'Seurat' or 'SingleCellExperiment' object.
}
