% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPseudobulk.R
\name{getPseudobulk}
\alias{getPseudobulk}
\title{Calculate pseudobulk values}
\usage{
getPseudobulk(
  object,
  replicate_labels,
  split_labels = NULL,
  use_cells = NULL,
  min_cells_per_split = 100,
  min_cells_per_replicate = 10,
  min_replicates_per_split = 6,
  min_cells_per_feature = 10,
  min_prop_cells_per_feature = 0.1,
  use_assay = NULL,
  use_layer = NULL,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{Seurat}, \code{SingleCellExperiment},
or \code{matrix}. Data supplied as class \code{matrix} should be a
feature x cell matrix.}

\item{replicate_labels}{A string indicating the name of the
metadata column containing the biological replicate labels or a character
vector containing the biological replicate labels in order. The biological
replicate labels are used to construct/define the pseudobulks.}

\item{split_labels}{A string indicating the name of a
column by which to split the cells prior to pseudobulking and performing
differential expression (e.g., cell types). Alternately, a character vector
containing the split labels for each cell in order. Results will be returned
for each unique value indicated by \code{split_labels}. Default = \code{NULL}
will generate a single pseudobulk matrix that includes all cells.}

\item{use_cells}{A vector of cell names to subset the object to prior to
subsequent pseudobulk steps. Default = \code{NULL} will use all cells.}

\item{min_cells_per_split}{A numeric value indicating the minimum number of
cells within one split. Pseudobulk steps will not be performed for splits
with fewer cells. Defaults to 100.}

\item{min_cells_per_replicate}{A numeric value indicating the minimum number
of cells within one replicate for one split. Pseudobulk steps will not be
performed for replicates with fewer cells for that split. Defaults to 10.}

\item{min_replicates_per_split}{A numeric value indicating the minimum number
of distinct replicates represented within one split. Pseudobulk steps
will not be performed for splits with fewer replicates. Defaults to 6.}

\item{min_cells_per_feature}{A numeric value indicating the minimum number
of cells (within a split) with expression of a feature. Pseudobulk expression
will not be calculated for genes expressed in fewer cells. Defaults to 10.}

\item{min_prop_cells_per_feature}{A numeric value indicating the minimum
proportion of cells (within a split) with expression of a feature. Pseudobulk
expression will not be calculated for features expressed in fewer cells.
Defaults to 0.1.}

\item{use_assay}{A string indicating the assay to use in the
provided object. Default = \code{NULL} will choose the current active assay
for \code{Seurat} objects and the \code{counts} assay for
\code{SingleCellExperiment} objects.}

\item{use_layer}{For \code{Seurat} objects, a string or vector
indicating the layer—previously known as slot—to use in the provided object.
Default = \code{NULL} will use the \code{counts} layer.}

\item{n_cores}{A numeric value indicating the number of cores to use for
parallelization. Default = \code{NULL} will use the number of available cores
minus 2.}

\item{verbose}{A Boolean value indicating whether to use verbose output
during the execution of this function. Defaults to \code{TRUE}.
Can be set to \code{FALSE} for a cleaner output.}
}
\value{
Returns a list containing the following elements: \describe{
\item{PB_values}{A list of feature x replicate matri(ces) containing
pseudobulk values for each feature, one matrix per split}
\item{metadata}{Dataframe record of quality control metrics for each split}
\item{parameters}{List recording parameter values used}
}
}
\description{
This function will generate pseudobulk expression matri(ces) from a provided
Seurat object, SingleCellExperiment object, or feature x cell matrix.
Multiple pseudobulk matrices can be generated across different divisions of
a dataset, such as cell types.
}
