% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingUtils.R
\name{permuteDEpalette}
\alias{permuteDEpalette}
\title{Generate color palette}
\usage{
permuteDEpalette(type = "discrete", n = NULL, palette = NULL, swatch = FALSE)
}
\arguments{
\item{type}{A character string indicating the palette type. Permitted values
are "discrete" and "gradient". Defaults to discrete.}

\item{n}{Number of colors. Default = \code{NULL} will return all of the
pre-set colors in the palette.}

\item{palette}{A character string indicating the palette name. Permitted
values are "choir", "archr", "inferno", and "frozen". Default =
\code{NULL} will use "choir" for discrete colors and "inferno" for
gradient colors. Alternately, provide a vector of color values to use as
starting values for the color palette.}

\item{swatch}{A Boolean value indicating whether to plot a swatch of the
palette.}
}
\value{
Returns a vector of n hex values.
}
\description{
Generate a color palette. For discrete palettes, hex values are standard up
to n = 100, but for larger values of n, they are generated using
\code{Polychrome::createPalette()}.
}
\details{
The "choir" palette is adapted from function  \code{CHOIR::CHOIRpalette}
in R package \code{CHOIR} (Sant et al. 2025). The "archr" palette is
adapted from palette "stallion" in R package \code{ArchR} (Granja & Corces
et al. 2020).
}
