% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingUtils.R
\name{plotFeature}
\alias{plotFeature}
\title{Plot levels of a single feature across groups}
\usage{
plotFeature(
  input,
  feature_name,
  use_splits = NULL,
  normalization_method = "cpm",
  plot_type = "boxplot",
  title = NULL,
  subtitle = NULL,
  label_replicates = FALSE,
  label_statistics = TRUE
)
}
\arguments{
\item{input}{Output from function \code{runDE}.}

\item{feature_name}{A character string indicating which feature to plot.}

\item{use_splits}{A character string or vector containing the names of splits
to use. Defaults to \code{NULL}, which will try all splits.}

\item{normalization_method}{A character string indicating which normalization
method to apply to the replicate x feature matrix. Permitted values are
"cpm", "log_cpm", and "none". Defaults to "cpm".}

\item{plot_type}{A character string indicating the type of plot. Permitted
values are "boxplot", "bar_se", "bar_sd", "beeswarm".
Defaults to "boxplot".}

\item{title}{A character string indicating the plot title. Default =
\code{NULL} sets a title automatically for each split.}

\item{subtitle}{Character string indicating the plot subtitle. Default =
\code{NULL} automatically generates a subtitle.}

\item{label_replicates}{A Boolean value indicating whether to label the
replicates. Defaults to \code{FALSE}.}

\item{label_statistics}{A Boolean value indicating whether to label the
differential expression analysis LFC and p-value. Defaults to TRUE.}
}
\value{
If only one split is provided, a single \code{ggplot2} object,
otherwise a named list of \code{ggplot2} objects, where each element
corresponds to a split.
}
\description{
This functions takes the output from function \code{runDE} and plots the
expression of a single feature across the replicates in each group for each
indicated split.
}
