% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDE.R
\name{runDE}
\alias{runDE}
\title{Run differential expression analysis to compare two groups}
\usage{
runDE(
  object,
  replicate_labels = NULL,
  group_labels,
  split_labels = NULL,
  reference_group = NULL,
  use_cells = NULL,
  pseudobulk = "generate",
  de_method = "edgeR",
  de_test = "LRT",
  de_params = list(),
  p_adjust_method = "fdr",
  min_cells_per_split = 100,
  min_cells_per_replicate = 10,
  min_replicates_per_split = 6,
  min_replicates_per_group = 3,
  min_cells_per_feature = 10,
  min_prop_cells_per_feature = 0.1,
  force_balance = FALSE,
  use_assay = NULL,
  use_layer = NULL,
  random_seed = 1,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{Seurat}, \code{SingleCellExperiment},
or \code{matrix}. Data supplied as class \code{matrix} may be either a
feature x cell matrix or a pre-computed pseudobulk feature x replicate
matrix. Note that raw counts are expected, and the normalization method
applied during differential expression analysis differs across the methods
and tests.}

\item{replicate_labels}{A string indicating the name of the
metadata column containing the biological replicate labels or a character
vector containing the biological replicate labels in order. For pseudobulk DE
analysis, the biological replicate labels are used to construct/define the
pseudobulks. For cell-level DE analysis, the biological replicate labels are
not used in this function, but will be passed on to function permuteDE so
that the permutation of group labels keeps replicates intact.}

\item{group_labels}{A string indicating the name of the
column containing the two comparison group labels or a character vector
containing the comparison labels in order.}

\item{split_labels}{A string indicating the name of a
column by which to split the cells prior to pseudobulking and performing
differential expression (e.g., cell types). Alternately, a character vector
containing the split labels for each cell in order. Results will be returned
for each unique value indicated by \code{split_labels}. Default = \code{NULL}
will run pseudobulk differential expression on all cells together.}

\item{reference_group}{A string specifying the reference group. Defaults to
\code{NULL}, in which case the first value alphabetically is used as the
reference.}

\item{use_cells}{A vector of cell names to subset the object to prior to
subsequent pseudobulk and differential expression steps. Default =
\code{NULL} will use all cells.}

\item{pseudobulk}{A string indicating pseudobulk handling. Permitted values
are: "generate" (pseudobulk matrices will be generated), "supplied"
(pseudobulk matrix was supplied by the user to parameter \code{object}), or
"none" (pseudobulking will not be used, cell-level differential expression
analysis will be run). Defaults to "generate".}

\item{de_method}{Which tool to use for differential expression analysis.
Permitted values are "edgeR", "DESeq2", "limma", and "presto". Defaults to
"edgeR".}

\item{de_test}{Which test to use for differential expression analysis.
Available values are dependent on the \code{de_method}: "edgeR" ("LRT",
"QLF", "exact"), "DESeq2" ("LRT", "Wald"), "limma" ("trend", "voom"),
"presto" ("wilcox_cpm", "wilcox_log_cpm"). Defaults to "LRT".}

\item{de_params}{A list of lists containing additional parameters to be
passed to specific DE functions. The name of each element must be the
specific DE function to which those parameters are passed. Defaults to an
empty list.}

\item{p_adjust_method}{A string indicating which multiple comparison
adjustment to use. For permitted values, see \code{stats::p.adjust.methods}.
Defaults to "fdr" (Benjamini & Hochberg, 1995).}

\item{min_cells_per_split}{A numeric value indicating the minimum number of
cells within one split. Pseudobulk and differential expression steps will not
be performed for splits with fewer cells. Defaults to 100.}

\item{min_cells_per_replicate}{A numeric value indicating the minimum number
of cells within one replicate for one split. Pseudobulk steps will not be
performed for replicates with fewer cells for that split. Defaults to 10.}

\item{min_replicates_per_split}{A numeric value indicating the minimum number
of distinct replicates represented within one split. Pseudobulk expression
and differential expression will not be performed for splits with fewer
replicates. Defaults to 6.}

\item{min_replicates_per_group}{A numeric value indicating the minimum number
of distinct replicates represented within each of the two comparison groups.
Pseudobulk and differential expression steps will not be performed for
splits with fewer replicates. Defaults to 3.}

\item{min_cells_per_feature}{A numeric value indicating the minimum number
of cells (within a split) with expression of a gene. Pseudobulk and
differential expression will not be calculated for genes expressed in
fewer cells. Defaults to 10.}

\item{min_prop_cells_per_feature}{A numeric value indicating the minimum
proportion of cells (within a split) with expression of a gene. Pseudobulk
and differential expression will not be calculated for genes
expressed in fewer cells. Defaults to 0.1.}

\item{force_balance}{A boolean indicating whether to force the two comparison
groups to have the same sample size. Defaults to \code{FALSE}. If
\code{TRUE}, the larger group will be randomly downsampled to the size of the
smaller group.}

\item{use_assay}{A string indicating the assay to use in the
provided object. Default = \code{NULL} will choose the current active assay
for \code{Seurat} objects and the \code{counts} assay for
\code{SingleCellExperiment} objects.}

\item{use_layer}{For \code{Seurat} objects, a string or vector
indicating the layer—previously known as slot—to use in the provided object.
Default = \code{NULL} will use the \code{counts} layer.}

\item{random_seed}{A numerical value indicating the random seed to be used.
Defaults to 1. Only relevant in this function when parameter
\code{force_balance = TRUE}.}

\item{n_cores}{A numeric value indicating the number of cores to use for
parallelization. Default = \code{NULL} will use the number of available cores
minus 2.}

\item{verbose}{A Boolean value indicating whether to use verbose output
during the execution of this function. Defaults to \code{TRUE}.
Can be set to \code{FALSE} for a cleaner output.}
}
\value{
Returns a list containing the following elements: \describe{
\item{DE_results}{Dataframe containing DE results for each feature, by
split}
\item{PB_values}{If using pseudobulk data, a list of feature x replicate
matri(ces) containing pseudobulk values for each feature, one matrix per
split}
\item{cell_values}{Alternately, if using cell-level data, a list of feature x
cell matri(ces) containing counts for each feature, one matrix per
split}
\item{metadata}{List recording characteristics of the data and runtime}
\item{parameters}{List recording parameter values used}
}
}
\description{
This function identifies differentially expressed features between two groups
using indicated differential expression analysis methods.
}
\details{
By default, pseudobulk matri(ces) are generated or supplied by the user, then
used to run pseudobulk differential expression. The following existing tools
are supported: \code{edgeR}, \code{DESeq2}, \code{limma}, and the Wilcoxon
rank-sum test. Alternately, users may skip pseudobulking and run cell-level
differential expression (not recommended in most cases).
}
